+++ tests/test_config_state.py
@@ -50,6 +50,9 @@ def test_load_wrapper_config_autodetect_regional(tmp_working_dir_regional, defau
                     region: eu-west-1
                 credentials:
                     profile: terraform-states-profile # the AWS profile to use for state storage
+
+            backend_parameters:
+                state_snaphot: "false"
             """
         )
     )
@@ -67,6 +70,9 @@ def test_load_wrapper_config_autodetect_regional(tmp_working_dir_regional, defau
     assert wrapper_config["state"]["aws"]["state_account"] == "12345678910"
     assert wrapper_config["state"]["aws"]["state_region"] == "eu-west-1"
     assert wrapper_config["state"]["aws"]["state_profile"] == "terraform-states-profile"
+    assert wrapper_config["state"]["aws"]["account"] == "12345678910"
+    assert wrapper_config["state"]["aws"]["region"] == "eu-west-1"
+    assert wrapper_config["state"]["aws"]["state_backend_parameters"]["state_snaphot"] == "false"
 
 
 def test_load_wrapper_config_autodetect_global(tmp_working_dir_global, default_args):
@@ -81,6 +87,9 @@ def test_load_wrapper_config_autodetect_global(tmp_working_dir_global, default_a
                     account: '12345678910' # the AWS account to use for state storage
                 credentials:
                     profile: terraform-states-profile # the AWS profile to use for state storage
+
+            backend_parameters:
+                state_snaphot: "false"
             """
         )
     )
@@ -97,6 +106,8 @@ def test_load_wrapper_config_autodetect_global(tmp_working_dir_global, default_a
     assert wrapper_config["stack"] == "teststack"
     assert wrapper_config["state"]["aws"]["state_account"] == "12345678910"
     assert wrapper_config["state"]["aws"]["state_profile"] == "terraform-states-profile"
+    assert wrapper_config["state"]["aws"]["account"] == "12345678910"
+    assert wrapper_config["state"]["aws"]["state_backend_parameters"]["state_snaphot"] == "false"
 
 
 def test_load_wrapper_config_list(tmp_working_dir_regional, default_args):
@@ -118,6 +129,9 @@ def test_load_wrapper_config_list(tmp_working_dir_regional, default_args):
                       region: eu-west-2
                   credentials:
                       profile: terraform-states-profile2 # the AWS profile to use for state storage
+
+            backend_parameters:
+                state_snaphot: "false"
             """
         )
     )
@@ -132,12 +146,18 @@ def test_load_wrapper_config_list(tmp_working_dir_regional, default_args):
     assert wrapper_config["environment"] == "testenvironment"
     assert wrapper_config["region"] == "testregion"
     assert wrapper_config["stack"] == "teststack"
+    assert wrapper_config["state"]["aws"]["account"] == "12345678910"
+    assert wrapper_config["state"]["aws"]["region"] == "eu-west-1"
     assert wrapper_config["state"]["aws"]["state_account"] == "12345678910"
     assert wrapper_config["state"]["aws"]["state_region"] == "eu-west-1"
     assert wrapper_config["state"]["aws"]["state_profile"] == "terraform-states-profile"
+    assert wrapper_config["state"]["aws"]["state_backend_parameters"]["state_snaphot"] == "false"
+    assert wrapper_config["state"]["aws_backend2"]["account"] == "10987654321"
+    assert wrapper_config["state"]["aws_backend2"]["region"] == "eu-west-2"
     assert wrapper_config["state"]["aws_backend2"]["state_account"] == "10987654321"
     assert wrapper_config["state"]["aws_backend2"]["state_region"] == "eu-west-2"
     assert wrapper_config["state"]["aws_backend2"]["state_profile"] == "terraform-states-profile2"
+    assert wrapper_config["state"]["aws_backend2"]["state_backend_parameters"]["state_snaphot"] == "false"
 
 
 def test_load_wrapper_config_old_map(tmp_working_dir_regional, default_args):
@@ -153,12 +173,9 @@ def test_load_wrapper_config_old_map(tmp_working_dir_regional, default_args):
                     region: eu-west-1
                 credentials:
                     profile: terraform-states-profile # the AWS profile to use for state storage
-            azure:
-                name: azure_backend
-                general:
-                    subscription_uid: 'f6eaa78a-3adf-4818-b0ea-87fb07acf477'
-                    resource_group_name: my-resource-group
-                    storage_account_name: mystorage
+
+            backend_parameters:
+                state_snaphot: "false"
             """
         )
     )
@@ -176,6 +193,6 @@ def test_load_wrapper_config_old_map(tmp_working_dir_regional, default_args):
     assert wrapper_config["state"]["aws"]["state_account"] == "12345678910"
     assert wrapper_config["state"]["aws"]["state_region"] == "eu-west-1"
     assert wrapper_config["state"]["aws"]["state_profile"] == "terraform-states-profile"
-    assert wrapper_config["state"]["azure_backend"]["state_subscription"] == "f6eaa78a-3adf-4818-b0ea-87fb07acf477"
-    assert wrapper_config["state"]["azure_backend"]["state_rg"] == "my-resource-group"
-    assert wrapper_config["state"]["azure_backend"]["state_storage"] == "mystorage"
+    assert wrapper_config["state"]["aws"]["account"] == "12345678910"
+    assert wrapper_config["state"]["aws"]["region"] == "eu-west-1"
+    assert wrapper_config["state"]["aws"]["state_backend_parameters"]["state_snaphot"] == "false"